rename custom selector model package name to customselector
