+++ app/src/main/java/fr/free/nrw/commons/customselector/listeners/ImageSelectListener.kt
@@ -1,7 +1,7 @@
 package fr.free.nrw.commons.customselector.listeners
 
 import android.net.Uri
-import fr.free.nrw.commons.data.models.model.Image
+import fr.free.nrw.commons.data.models.customselector.Image
 
 /**
  * Custom selector Image select listener
