+++ app/src/main/java/fr/free/nrw/commons/customselector/ui/adapter/FolderAdapter.kt
@@ -10,8 +10,8 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import fr.free.nrw.commons.R
 import fr.free.nrw.commons.customselector.listeners.FolderClickListener
-import fr.free.nrw.commons.data.models.model.Folder
-import fr.free.nrw.commons.data.models.model.Image
+import fr.free.nrw.commons.data.models.customselector.Folder
+import fr.free.nrw.commons.data.models.customselector.Image
 
 /**
  * Custom selector FolderAdapter.
