+++ app/src/main/java/fr/free/nrw/commons/customselector/ui/selector/CustomSelectorViewModel.kt
@@ -4,9 +4,9 @@ import android.content.Context
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import fr.free.nrw.commons.customselector.listeners.ImageLoaderListener
-import fr.free.nrw.commons.data.models.model.CallbackStatus
-import fr.free.nrw.commons.data.models.model.Image
-import fr.free.nrw.commons.data.models.model.Result
+import fr.free.nrw.commons.data.models.customselector.CallbackStatus
+import fr.free.nrw.commons.data.models.customselector.Image
+import fr.free.nrw.commons.data.models.customselector.Result
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.cancel
