+++ app/src/main/java/fr/free/nrw/commons/data/models/customselector/Image.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.data.models.model
+package fr.free.nrw.commons.data.models.customselector
 
 import android.net.Uri
 import android.os.Parcel
