+++ app/src/main/java/fr/free/nrw/commons/data/models/customselector/Result.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.data.models.model
+package fr.free.nrw.commons.data.models.customselector
 
 /**
  * Custom selector data class Result.
@@ -7,7 +7,7 @@ data class Result(
     /**
      * CallbackStatus : stores the result status
      */
-    val status:CallbackStatus,
+    val status: CallbackStatus,
 
     /**
      * Images : images retrieved
